Add files via upload
This project presents a Power BI project that analyzes pizza order and sales data, providing insights into various aspects of the business. It utilizes an interactive dashboard for visual analysis of pizza order and sales data. The project employs Power BI's data visualization capabilities to create an interactive and dynamic dashboard that provides insights into various aspects of the business.